fix: Keep configs without a header first in struct
It was not possible to write config that contains both search_dirs and
default_session for example. This was due to search_dirs use headers
and the default_session was written after search_dirs, which resulted in
default_session ending up below one of those headers. This resulted in
an error when writing to file.