create tmux struct and remove execute_tmux_command function
This is the first step to making this crate more testable.
In order to test this crate, we need to be able to execute tmux
commands against a tmux socket that is not the one we are currently in.

Currently, `execute_tmux_command` is used to interact with tmux,
and this makes it difficult to test because you would have to manually
add `-L test-socket` to each invocation of that method.

By creating a tmux struct and storing the socket name in there, it
becomes possible to test.