Add support for running script on session creation
Looks for a script named .tms-create in the path of the session being
created. Runs it by sending keys to the first window's top pane.
The script needs to be executable.

This can be overridden by adding something like the following to the
config file:
[session-configs.session_name]
create_script = "my-create.sh"

fixes #33