Add config option for listing worktrees for bare repositories
Note that for git repos, this currently undoes most of the lazy repos
optimization, since all repos have to be opened with gix again in order
to retrieve their worktrees.
If performance turns out to be a problem, then we could instead see if
reading worktrees from $GIT_DIR/worktrees works better. But since this
option is opt-in, and performance was not a concern for all usecases
prior to this optimization, it's probably fine to keep a (possibly
noticeable, probably not) slower startup for those who prefer to have
this option enabled.
For jj it's not as much of a concern, but the decoding of the workspaces
had to be duplicated from jj's internals, which also necessitates the
prost crate addition.

So just as before, the user's VCS providers priority config and the
amount of repos of each determines the impact enabling this setting will
have on performance.

Also adds a check to set_up_tmux_env so that when opening a worktree
directly, the main repo's other worktrees won't be opened as
windows. Since we're opening one directly, the assumption is that we
specificially want only the one selected.

Co-authored-by: Swann HERRERA <swann.herrera@nuant.com>