Add rust-toolchain.toml + update flake.nix (#155)
* fix clippy lints

* add rust-toolchain.toml

CI broke because of new clippy lints, so we should probably lock the
version.
Also reworks the nix flake to no longer apply overrides to the upstream
nixpkgs package, because users using the overlay might not all apply it
to the same nixpkgs version and break compatibility, like when we changed
the shell completion generations.
The crane package also has the benefits of being useful for CI due to
including all checks (clippy, format, tests)

* actions: add nix workflow