Add support for jujutsu (#160)
uses jj-lib crate

This adds a config option `vcs_providers`, where both `git` and `jj`
can be added. By default only git is enabled.
Providers are used lazily in order, so if jujutsu is configured before git,
git will not find a colocated jj repository if jj found it first.

Worktrees/workspaces are handled differently depending on vcs provider.
Git should behave the same as before, with the exception that on non-bare
repos worktrees will be opened as new windows as well.
jj doesn't have bare repos, and so far there doesn't seem to be a reliable
way to refer from the "base" repo to its workspaces, so instead tms will
scan all configured search paths for workspaces belonging to the base
workspace that is opened, and then open a new window for each of them.
(jj workspaces are identified by `.jj/repo` not being the actual repo,
but instead a file that points to the main workspace where the repo is
stored)