clone/init repo: skip search paths with depth == 0 (#167)
If a search path is configured with depth 0, its subdirectories will not
be searched, so cloning into it is pointless. The only reason to have it
configured if there is a singular repository in an isolated location
which should show up on the picker, and in that case offering it as an
option does not make sense.